chore: fix warnings for ci

+++ src/indexer/evm/mod.rs
@@ -48,7 +48,7 @@ impl ChainIndexer for EVMIndexer {
                         Ok(parsed) => {
                             let last_synced = last_synced::ActiveModel {
                                 chain_id: Set(id as i64),
-                                block_number: Set(parsed.block_number as i64),
+                                block_number: Set(parsed.block_number),
                             };
                             if let Err(e) =
                                 db::insert_model(parsed.model, last_synced, &live_indexer.db).await
@@ -93,7 +93,7 @@ impl EVMIndexer {
                 Ok(parsed) => {
                     let last_synced = last_synced::ActiveModel {
                         chain_id: Set(id as i64),
-                        block_number: Set(parsed.block_number as i64),
+                        block_number: Set(parsed.block_number),
                     };
                     db::insert_model(parsed.model, last_synced, &self.db).await?;
                 }
